samples/python/agents/a2a_help/requirements.txt
google-adk[a2a] @ git+https://github.com/google/adk-python.git@main
a2a-sdk>=0.3.0
